Refactor project

+++ game/Types.h
@@ -7,7 +7,7 @@
 
 // Source: https://gist.github.com/Lee-R/3839813
 constexpr auto fnv1a_32(char const *s, const std::size_t count) -> std::uint32_t { // NOLINT(*-no-recursion)
-    return count ? (fnv1a_32(s, count - 1) ^ s[count - 1]) * 16777619u : 2166136261u;
+    return count ? (fnv1a_32(s, count - 1) ^ static_cast<std::uint32_t>(s[count - 1])) * 16777619u : 2166136261u;
 }
 
 constexpr std::uint32_t operator "" _hash(char const *s, const std::size_t count) {
